Split up text extension methods.
